+++ .gitignore
@@ -1,3 +1,3 @@
 # 패키지 파일 무시
-node_modules/
+node_modules
 .env
