+++ data/var/lib/mysql/ib_buffer_pool
@@ -1,6 +1,4 @@
 7,3
-5,3
-4,3
 1,45
 3,44
 2,44
@@ -122,23 +120,37 @@
 3,5
 2,5
 1,5
-3,4
-2,4
-1,4
-3,3
-2,3
 1,3
 0,6
+0,0
 0,47
 0,46
 0,49
 0,48
 0,45
+8,3
+8,2
+8,1
+5,3
+4,3
+3,4
+3,3
+3,2
+3,0
+2,4
+2,3
+2,2
+2,0
+1,4
+1,2
+1,0
 0,12
-0,10
-0,8
 0,11
+0,10
+0,9
+0,8
 0,5
 0,7
 0,4
 0,3
+8,0
