+++ models/Users.js
@@ -1,0 +1,24 @@
+const { Sequelize, DataTypes } = require('sequelize');
+const db = require('../config/db');
+
+// User 모델 정의
+const Users = db.sequelize.define('Users', {
+  id: {
+    type: DataTypes.INTEGER,
+    autoIncrement: true,
+    primaryKey: true
+  },
+  username: {
+    type: DataTypes.STRING,
+    allowNull: false
+  },
+  email: {
+    type: DataTypes.STRING,
+    allowNull: false
+  }
+}, {
+  tableName: 'Users',
+  timestamps: false
+});
+
+module.exports = Users;
